+++ contextualized_topic_models/__init__.py
@@ -2,4 +2,6 @@
 
 __author__ = """Federico Bianchi"""
 __email__ = 'f.bianchi@unibocconi.it'
-__version__ = '1.8.2'
+__version__ = '2.0.0'
+
+