+++ contextualized_topic_models/utils/early_stopping/early_stopping.py
@@ -39,6 +39,7 @@ class EarlyStopping:
         if self.best_score is None:
             self.best_score = score
             self.save_checkpoint(val_loss, model)
+            model.best_components = model.model.beta.clone()
         elif score < self.best_score + self.delta:
             self.counter += 1
             self.trace_func(f'EarlyStopping counter: {self.counter} out of {self.patience}')
@@ -46,6 +47,7 @@ class EarlyStopping:
                 self.early_stop = True
         else:
             self.best_score = score
+            model.best_components = model.model.beta.clone()
             self.save_checkpoint(val_loss, model)
             self.counter = 0